add access control for deletion

+++ grails-app/controllers/au/org/ala/collectory/DataProviderController.groovy
@@ -7,6 +7,7 @@ class DataProviderController extends ProviderGroupController {
         entityNameLower = "dataProvider"
     }
 
+
     def index = {
         redirect(action:"list")
     }
@@ -38,27 +39,31 @@ class DataProviderController extends ProviderGroupController {
     def delete = {
         def instance = get(params.id)
         if (instance) {
-            /* need to remove it as a parent from all children otherwise they will be deleted */
-            def resources = instance.resources as List
-            resources.each {
-                instance.removeFromResources it
-                it.userLastModified = username()
-                it.save()  // necessary?
-            }
-            // remove contact links (does not remove the contact)
-            ContactFor.findAllByEntityUid(instance.uid).each {
-                it.delete()
-            }
-            // now delete
-            try {
-                ActivityLog.log username(), isAdmin(), params.id as long, Action.DELETE
-                instance.delete(flush: true)
-                flash.message = "${message(code: 'default.deleted.message', args: [message(code: 'dataProvider.label', default: 'dataProvider'), params.id])}"
-                redirect(action: "list")
-            }
-            catch (org.springframework.dao.DataIntegrityViolationException e) {
-                flash.message = "${message(code: 'default.not.deleted.message', args: [message(code: 'dataProvider.label', default: 'dataProvider'), params.id])}"
-                redirect(action: "show", id: params.id)
+            if (isAdmin()) {
+                /* need to remove it as a parent from all children otherwise they will be deleted */
+                def resources = instance.resources as List
+                resources.each {
+                    instance.removeFromResources it
+                    it.userLastModified = username()
+                    it.save()  // necessary?
+                }
+                // remove contact links (does not remove the contact)
+                ContactFor.findAllByEntityUid(instance.uid).each {
+                    it.delete()
+                }
+                // now delete
+                try {
+                    ActivityLog.log username(), isAdmin(), params.id as long, Action.DELETE
+                    instance.delete(flush: true)
+                    flash.message = "${message(code: 'default.deleted.message', args: [message(code: 'dataProvider.label', default: 'dataProvider'), params.id])}"
+                    redirect(action: "list")
+                }
+                catch (org.springframework.dao.DataIntegrityViolationException e) {
+                    flash.message = "${message(code: 'default.not.deleted.message', args: [message(code: 'dataProvider.label', default: 'dataProvider'), params.id])}"
+                    redirect(action: "show", id: params.id)
+                }
+            } else {
+                render("You are not authorised to access this page.")
             }
         }
         else {
